% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_stimulus_cache.R
\name{clear_stimulus_cache}
\alias{clear_stimulus_cache}
\title{Clear cache used by stimulus.plot() to re-use resampling results from identical previous calls.}
\usage{
clear_stimulus_cache()
}
\description{
When stimulus.plot() type='effects' is run, the resamples are saved to the local environment and if
an identical call is ran again (same data, same dv, same condition, same number of simulations, etc)
the results are loaded instead of re-calculated. You can force recalculation by deleting the cache with this function.
}
