% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimulus.beeswarm.R
\name{stimulus.beeswarm}
\alias{stimulus.beeswarm}
\title{Make beeswarm plots for stimuli in compared-stimulus design (see "Stimulus Sampling Reimagined" by Simonsohn, Montealegre, & Evangelidis (2024))}
\usage{
stimulus.beeswarm(
  data,
  dv,
  stimulus,
  condition,
  flip.conditions = FALSE,
  dv.is.percentage = FALSE,
  simtot = 500,
  confidence = 95,
  ylim = c(),
  ylab1 = "",
  ylab2 = "",
  xlab1 = "",
  xlab2 = NULL,
  dot.spacing = "auto",
  col1 = "blue4",
  col2 = "red4",
  main = "",
  watermark = TRUE,
  save.as = "",
  svg.width = "",
  svg.height = "",
  ...
)
}
\arguments{
\item{data}{dataframe containing variables to be analyzed}

\item{dv}{name of the dependent variable (e.g., dv='y'), quotes are not required}

\item{stimulus}{name of the variable containing the stimulus ID
(e.g., stimulus='stim_id'), quotes are not required. The 'markers' in the beeswarm
will be the text for this labels, if you want the markers to be different
(e.g., abbreviations) or single word summaries, create a new variable with those
values and use that in the stimulus.beeswarm() call}

\item{condition}{name of the variable containing the condition indicator
(e.g., condition='cond'), quotes are not required. The x-axis label will use values
in this variable to identify the two conditions. If you want to customize the legend
(e.g., so that it is not "0" vs "1"), create new variable, say df$cond2
with descriptive values for condition (e.g., df$cond2=ifelse(df$cond==1,'control','treatment'))
and then use condition='cond2' in the stimulus.plot call.}

\item{flip.conditions}{by default the condition labels are sorted alphabetically.
Set flip.conditions=TRUE to reverse the order}

\item{dv.is.percentage}{if set to TRUE values of the dependent variable are formatted as percentages}

\item{simtot}{integer value for number of bootstraps used to build a confidence band
for the range of means under homogeneity}

\item{confidence}{integer value for confidence level for confidence band. Defaults to 95.}

\item{ylab1, ylab2}{labels on the y-axis (optional)}

\item{dot.spacing}{scalar for horizontal distance between labels with stimuli names}

\item{col1, col2}{strings with colors for first and second condition in the plot
(lighter versions of colors are automatically generated)}

\item{main}{figure name, same as base-R plot main argument}

\item{watermark}{set to FALSE to not display {stimulus version} in bottom left of figure}

\item{save.as}{filepath for saving figure. Must be .svg or .png file (optional)}
}
\description{
Make beeswarm plots for stimuli in compared-stimulus design (see "Stimulus Sampling Reimagined" by Simonsohn, Montealegre, & Evangelidis (2024))
}
