% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimulus beeswarm.R
\name{stimulus.beeswarm}
\alias{stimulus.beeswarm}
\title{Make beeswarm plots for stimuli in compared-stimulus design (see "Stimulus Sampling Reimagined" by Simonsohn, Montealegre, & Evangelidis (2024))}
\usage{
stimulus.beeswarm(
  data,
  dv,
  stimulus,
  condition,
  flip.conditions = FALSE,
  dv.is.percentage = FALSE,
  ylab1 = "",
  ylab2 = "",
  xlab1 = "",
  xlab2 = NULL,
  ylim = c(),
  dot.spacing = "auto",
  col1 = "blue4",
  col2 = "red4",
  main = "",
  watermark = TRUE,
  save.as = "",
  svg.width = "",
  svg.height = "",
  ...
)
}
\arguments{
\item{data}{dataframe containing variables to be analyzed}

\item{dv}{name of the dependent variable (e.g., dv='y'), quotes are not required}

\item{stimulus}{name of the variable containing the stimulus ID
(e.g., stimulus='stim_id'), quotes are not required. The x-axis will have these values as tick-lables.
If you want to customize them, e.g., so that they have descriptive labels instead of numeric ID1, ID2...,
create a new variable with those descriptions and use that variable as the stimulus variable in the call
(e.g., df$item.id =c(1,2,3,4,5,...) --> df$item.id2=c('Chair','Airplane','Fries','Dalmatian','Coat',...))}

\item{condition}{name of the variable containing the condition indicator
(e.g., condition='cond'), quotes are not required. Figure legend will use values
in this variable to identify the two conditions. If you want to customize the legend
(e.g., so that it is not "0" vs "1"), create new variable, say df$cond2
with descriptive values for condition (e.g., df$cond2=ifelse(df$cond==1,'control','treatment'))
and then use condition='cond2' in the stimulus.plot call.}

\item{flip.conditions}{by default the condition labels are sorted alphabetically and the 2nd subtracted from the first.
set flip.conditions=TRUE so that the first is subtracted from the second instead}

\item{dv.is.percentage}{if set to TRUE values of the dependent variable are formatted as percentages}

\item{ylab1, ylab2}{labels on the y-axis (optional)}

\item{xlab1}{labels on the x-axis, centered, by default it is 'Condition' (optional)}

\item{xlab2}{vectro of size 2 with labels for the two conditions (optional)}

\item{dot.spacing}{horizontal distance between labels with stimuli names}

\item{main}{figure name, same as base-R plot main argument}

\item{watermark}{set to FALSE to not display {stimulus version} in bottom left of figure}

\item{save.as}{filepath for saving figure. Must be .svg or .png file (optional)}
}
\description{
Make beeswarm plots for stimuli in compared-stimulus design (see "Stimulus Sampling Reimagined" by Simonsohn, Montealegre, & Evangelidis (2024))
}
