% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimulus.plot.R
\name{stimulus.plot}
\alias{stimulus.plot}
\title{Make stimulus plots as in "Stimulus Sampling Reimagined", Simonsohn, Montealegre, & Evangelidis (2024)}
\usage{
stimulus.plot(
  data,
  dv,
  condition,
  stimulus,
  sort.by = "",
  plot.type = "means",
  flip.condition = FALSE,
  model = c(),
  overall.estimate = c(),
  overall.ci = c(),
  overall.p = c(),
  overall.label = c(),
  ylab1 = "",
  ylab2 = "",
  xlab1 = "Stimuli",
  xlab2 = "",
  cex = 1.5,
  value.labels.offset = -1,
  stimuli.numeric.labels = FALSE,
  label.low = "",
  label.high = "",
  decimals = 2,
  dv.is.percentage = FALSE,
  participant = "",
  legend.title = "",
  simtot = 100,
  watermark = TRUE,
  save.as = "",
  ...
)
}
\arguments{
\item{data}{dataframe containing variables to be analyzed}

\item{dv}{name of the dependent variable (e.g., dv='y'), quotes are not required}

\item{condition}{name of the variable containing the condition indicator (e.g., condition='treatment'), quotes are not required}

\item{stimulus}{name of the variable containing the stimulus ID (e.g., stimulus='stim_id'), quotes are not required}

\item{sort.by}{name of the variable to sort stimuli by in the plots. Defaults to sorting by observed effect size.
May be be useful to deviate to explore moderators or sort stimuli alphabetically.}

\item{plot.type}{can be either "means" or "effects", determines what's plotted in the y-axis of the figure}

\item{flip.condition}{whether to sort effect size in reverse order (TRUE/FALSE, defaults to FALSE)}

\item{ylab1, ylab2}{labels on the y-axis (character, optional)}

\item{xlab1, xlab2}{labels on the x-axis (character, optional)}

\item{decimals}{how many decimals to depict in the graph (integer)}

\item{participant}{name of the variable containing participant IDs; if set it's entered as random participant effect
(character, optional)}

\item{legend.title}{text above legend (character)}

\item{simtot}{number of simulations to rely on for estimating expected heterogeneity of
observed effect size. Only needed if plot.type='effects' (defaults to 100)}

\item{save.as}{name of file to save figure as (optional, filename must have extension .svg or .png)}

\item{supress.version.label}{set to TRUE to prevent version of {stimulus} package used
to generate figure from appearing in the bottom left of the figure}
}
\description{
Make stimulus plots as in "Stimulus Sampling Reimagined", Simonsohn, Montealegre, & Evangelidis (2024)
}
