% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimulus.plot.R
\name{stimulus.plot}
\alias{stimulus.plot}
\title{Make stimulus plots as in the article "Stimulus Sampling Reimagined" by Simonsohn, Montealegre, & Evangelidis (2024)}
\usage{
stimulus.plot(
  plot.type = "means",
  data,
  dv,
  condition,
  stimulus,
  participant = "",
  save.as = "",
  svg.width = "",
  svg.height = "",
  sort.by = "",
  flip.conditions = FALSE,
  model = c(),
  overall.estimate = c(),
  overall.ci = c(),
  overall.p = c(),
  overall.label = c(),
  ylab1 = "",
  ylab2 = "",
  xlab1 = "Stimuli",
  xlab2 = "",
  decimals = "auto",
  null.method = "shuffle",
  dv.is.percentage = FALSE,
  legend.title = "",
  simtot = 100,
  watermark = TRUE,
  seed = 2024,
  ylim = c(),
  ...
)
}
\arguments{
\item{plot.type}{can be either "means" or "effects", determines what's plotted in the y-axis of the figure}

\item{data}{dataframe containing variables to be analyzed}

\item{dv}{name of the dependent variable (e.g., dv='y'), quotes are not required}

\item{condition}{name of the variable containing the condition indicator
(e.g., condition='cond'), quotes are not required. Figure legend will use values
in this variable to identify the two conditions. If you want to customize the legend
(e.g., so that it is not "0" vs "1"), create new variable, say df$cond2
with descriptive values for condition (e.g., df$cond2=ifelse(df$cond==1,'control','treatment'))
and then use condition='cond2' in the stimulus.plot call.}

\item{stimulus}{name of the variable containing the stimulus ID
(e.g., stimulus='stim_id'), quotes are not required. The x-axis will have these values as tick-lables.
If you want to customize them, e.g., so that they have descriptive labels instead of numeric ID1, ID2...,
create a new variable with those descriptions and use that variable as the stimulus variable in the call
(e.g., df$item.id =c(1,2,3,4,5,...) --> df$item.id2=c('Chair','Airplane','Fries','Dalmatian','Coat',...))}

\item{participant}{name of the variable containing participant IDs; necessary for valid inference when
plot.type='effects' and each participant provided more than one observation}

\item{save.as}{filepath for saving figure. Must be .svg or .png file (optional)}

\item{sort.by}{name of variable to sort stimuli by. Defaults to sorting by observed effect size.}

\item{flip.conditions}{by default the condition labels are sorted alphabetically and the 2nd subtracted from the first.
set flip.conditions=TRUE so that the first is subtracted from the second instead}

\item{model}{method used to compute overall average: (1) 'regression', (2) mixed-model with stimulus intercepts,
and/or (3) mixed-model with random intercepts. If \code{participant} is provided, the regression clusters by participant
and the mixed models include participant random intercepts. Possible values: 'regression', 'intercepts', 'slopes', 'all'}

\item{overall.estimate}{scalar or vector of overall average effect if computed outside of \code{{stimulus}} (may not be set jointly with \code{model})}

\item{overall.ci}{vector of confidence interval bounds for \code{overall.estimate} (may not be set jointly with \code{model})}

\item{overall.p}{scalar or vector of p-value for overall average effect if computed outside of \code{{stimulus}} (may not be set jointly with \code{model})}

\item{overall.label}{label to show in x-axis for overall averages, can be used in conjuction with \code{model} to over-rule
the default labels of 'regression','Random Intercepts', and 'Random Slopes'}

\item{ylab1, ylab2}{labels on the y-axis (optional)}

\item{xlab1, xlab2}{labels on the x-axis (optional)}

\item{decimals}{force number of decimals to show for value labels (optional)}

\item{dv.is.percentage}{if set to TRUE values of the dependent variable are formatted as percentages}

\item{legend.title}{text with title above legend (optional)}

\item{simtot}{number of simulations to rely on for estimating expected heterogeneity of
observed effect size. Only needed if plot.type='effects' (defaults to 100)}

\item{watermark}{set to FALSE to not display {stimulus version} in bottom left of figure}

\item{seed}{used when resampling for the 'effects' plot, defaults to seed=2024}
}
\description{
Make stimulus plots as in the article "Stimulus Sampling Reimagined" by Simonsohn, Montealegre, & Evangelidis (2024)
}
