% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimulus.plot.R
\name{stimulus.plot}
\alias{stimulus.plot}
\title{Make stimulus plots as in Simonsohn, Montealegre, & Evangelidis (2024)}
\usage{
stimulus.plot(
  df,
  dv,
  condition,
  stimulus,
  plot.type = "means",
  flip.sign = FALSE,
  ylab1 = "",
  ylab2 = "",
  xlab1 = "Stimuli",
  xlab2 = "",
  value.labels.offset = -1,
  stimuli.numeric.labels = FALSE,
  label.low = "low",
  label.high = "high",
  decimals = 2,
  participant = "",
  legend.title = "",
  simtot = 100,
  ...
)
}
\arguments{
\item{df}{dataframe containing variables to be analyzied}

\item{dv}{the dependent variable (e.g., dv='y')}

\item{condition}{a string or numeric vector containing the condition indicator (e.g., condition='treatment')}

\item{stimulus}{a string or numeric vector identifying the different stimuli used}

\item{plot.type}{can be either "means" or "effects", determines what's plotted in the y-axis}

\item{flip.sign}{boolean on whether to sort effect size in reverse order (defaults to FALSE)}

\item{ylab1, ylab2}{string variables for labels on the y-axis (optional)}

\item{xlab1, xlab2}{string variables for labels on the x-axis (optional)}

\item{decimals}{numeric (integer), how many decimals to depict in the graph}
}
\description{
Make figure with stimulus level means or effects for experiments with multiple stimuli
}
