% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimulus.plot.R
\name{stimulus.plot}
\alias{stimulus.plot}
\title{Make stimulus plots as in "Stimulus Sampling Reimagined", Simonsohn, Montealegre, & Evangelidis (2024)}
\usage{
stimulus.plot(
  plot.type = "means",
  data,
  dv,
  condition,
  stimulus,
  participant = "",
  save.as = "",
  sort.by = "",
  flip.conditions = FALSE,
  model = c(),
  overall.estimate = c(),
  overall.ci = c(),
  overall.p = c(),
  overall.label = c(),
  ylab1 = "",
  ylab2 = "",
  xlab1 = "Stimuli",
  xlab2 = "",
  label.low = "",
  label.high = "",
  decimals = "auto",
  dv.is.percentage = FALSE,
  legend.title = "",
  simtot = 100,
  watermark = TRUE,
  seed = 2024,
  ...
)
}
\arguments{
\item{plot.type}{can be either "means" or "effects", determines what's plotted in the y-axis of the figure}

\item{data}{dataframe containing variables to be analyzed}

\item{dv}{name of the dependent variable (e.g., dv='y'), quotes are not required}

\item{condition}{name of the variable containing the condition indicator
(e.g., condition='treatment'), quotes are not required. Figure legend will use values
in 'condition' create new column with descriptive values if necessary
(e.g., cond=c(1,0),-->condition=c('Treatment','Controol'))}

\item{stimulus}{name of the variable containing the stimulus ID
(e.g., stimulus='stim_id'), quotes are not required. The x-axis labels will use values
in 'stimulus', create new column with descriptive values and use that in stimulus.plot() if desired
(e.g., df$item.id =c(1,2,3,4,5,...) --> df$item.id2=c('Chair','Airplace','Fries','Dalmatian','Coat',...))}

\item{participant}{name of the variable containing participant IDs; necessary for valid inference when
plot.type='effects' and each participant provided more than one observation}

\item{save.as}{filepath for saving figure. Must be .svg or .png file (optional)}

\item{sort.by}{name of variable to sort stimuli by. Defaults to sorting by observed effect size.}

\item{flip.conditions}{reverse order in which conditions are compared?
(e.g., treatment-control instead of control-treatment). Defaults to FALSE}

\item{model}{method used to compute overall average: (1) 'regression', (2) mixed-model with stimulus intercepts,
and/or (3) mixed-model with random intercepts. If \code{participant} is provided, the regression clusters by participant
and the mixed models include participant random intercepts. Possible values: 'regression', 'intercepts', 'slopes', 'all'}

\item{overall.estimate}{scalar or vector of overall average effect if computed outside of \code{{stimulus}} (may not be set jointly with \code{model})}

\item{overall.ci}{vector of confidence interval bounds for \code{overall.estimate} (may not be set jointly with \code{model})}

\item{overall.p}{scalar or vector of p-vauue for overall average effect if computed outside of \code{{stimulus}} (may not be set jointly with \code{model})}

\item{overall.label}{label to show in x-axis for overall averages, can be used in conjuction with \code{model} to over-rule
the default labels of 'regression','Random Intercepts', and 'Random Slopes'}

\item{ylab1, ylab2}{labels on the y-axis (optional)}

\item{xlab1, xlab2}{labels on the x-axis (optional)}

\item{decimals}{force number of decimals to show for value labels (optional)}

\item{dv.is.percentage}{if set to TRUE values are formatted as percentages}

\item{legend.title}{text above legend (optional)}

\item{simtot}{number of simulations to rely on for estimating expected heterogeneity of
observed effect size. Only needed if plot.type='effects' (defaults to 100)}

\item{watermark}{set to FALSE to not display {stimulus version} in bottom left of figure}

\item{seed}{used when resampling for the 'effects' plot, defaults to seed=2024}
}
\description{
Make stimulus plots as in "Stimulus Sampling Reimagined", Simonsohn, Montealegre, & Evangelidis (2024)
}
